Fixed destroy callbacks for contacts

+++ erp_base_erp_svcs/lib/erp_base_erp_svcs/extensions/active_record/has_contact.rb
@@ -39,7 +39,7 @@ module ErpBaseErpSvcs
 				  end
 
 				  def destroy_contact
-					  self.contact.destroy
+					  self.contact.destroy unless self.contact.nil?
 				  end
 
 				  def initialize_contact
